SPRK-PLATFORM-FOR-GITHUB/requirements.txt
# SPR{K}3 Platform - Python Dependencies

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# GitHub integration
PyGithub==2.1.1
cryptography==41.0.7
PyJWT==2.8.0

# Analysis engine
python-dateutil==2.8.2

# Background jobs (optional)
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
